[Feature] 회차별 정보 출력 기능 구현 및 DB 연동

+++ src/main/java/lotto/WebUILottoApplication.java
@@ -2,10 +2,10 @@ package lotto;
 
 import lotto.dao.GameResultDao;
 import lotto.dao.LottosDao;
+import lotto.dao.TurnDao;
 import lotto.dao.WinningLottoDao;
 import lotto.domain.*;
 import lotto.view.LottoDto;
-import lotto.view.LottosDto;
 import spark.ModelAndView;
 import spark.template.handlebars.HandlebarsTemplateEngine;
 
@@ -27,11 +27,26 @@ public class WebUILottoApplication {
         WinningLottoDao winningDao = new WinningLottoDao();
         GameResultDao resultDao = new GameResultDao();
         LottosDao lottosDao = new LottosDao();
+        TurnDao turnDao = new TurnDao();
 
         get("/", (req, res) -> {
             Map<String, Object> model = new HashMap<>();
             //TODO 횟수별 기능 구현
+            model.put("currentturn", turnDao.findNext());
+            model.put("turns", turnDao.findAll());
             return render(model, "form.html");
+        });
+
+        get("/turn", (req, res) -> {
+            Map<String, Object> model = new HashMap<>();
+            int turn = Integer.parseInt(req.queryParams("current_turn"));
+            GameResultDto result = resultDao.findByTurn(turn);
+            model.put("turn", turn);
+            model.put("lottos", lottosDao.findAllByTurn(turn));
+            model.put("winning_lotto", winningDao.findByTurn(turn));
+            model.put("result", parseResult2(result));
+            model.put("profit", result.getProfit());
+            return render(model, "turn_info.html");
         });
 
         post("/money", (req, res) -> {
@@ -56,11 +71,12 @@ public class WebUILottoApplication {
             Lotto lotto = parser.makeLotto(req.queryParams("lotto"));
             service.buy(lotto);
 
+            //자동 구매
             int autoCount = assignAutoPurchaseCount(service);
             model.put("manualCount", req.queryParams("manualCount"));
             model.put("autoCount", autoCount);
 
-            List<LottoDto> lottos = service.getLottos();
+            List<LottoDto> lottos = service.getLottos2();
             model.put("lottos", lottos);
 
             return render(model, "lottos.html");
@@ -73,23 +89,29 @@ public class WebUILottoApplication {
             Lotto lotto = parser.makeLotto(req.queryParams("winninglotto"));
             LottoNumber lottoNumber = parser.makeLottoNumber(Integer.parseInt(req.queryParams("bonusnumber")));
             WinningLotto winningLotto = WinningLotto.of(lotto, lottoNumber);
-            winningDao.add(winningLotto, 1);
+            winningDao.add(winningLotto, turnDao.findNext());
 
             LottoGameResult gameResult = service.gameResult();
             gameResult.match(winningLotto);
-            model.put("profit", String.format("%.1f", gameResult.profit(1000)));
-            model.put("stat", showGameResult(gameResult));
-            resultDao.add(new DtoConverter().convertResultToDto(gameResult),1);
+            model.put("profit", String.format("%.1f", gameResult.profit(LottoMachine.LOTTO_MONEY)));
+            model.put("stat", parseResult(gameResult));
+            resultDao.add(new DtoConverter().convertResultToDto(gameResult), turnDao.findNext());
 
             return render(model, "result.html");
         });
 
         post("/end", (req, res) -> {
             Map<String, Object> model = new HashMap<>();
+            if (req.queryParams("token").equals("restart")) {
+                turnDao.add();
+                model.put("currentturn", turnDao.findLast() + 1);
+                model.put("turns", turnDao.findAll());
+                return render(model, "form.html");
+            }
             resultDao.deleteAll();
             winningDao.deleteAll();
             lottosDao.deleteAll();
-
+            turnDao.deleteAll();
             return render(model, "form.html");
         });
     }
@@ -112,16 +134,24 @@ public class WebUILottoApplication {
         return autoPurchaseCount;
     }
 
-    private static List<String> showGameResult(final LottoGameResult gameResult) {
+    private static List<String> parseResult2(final GameResultDto result) {
         List<String> results = new ArrayList<>();
 
         for (Rank rank : Rank.reverseValues()) {
-            results.add(showRank(rank) + gameResult.getRankCount(rank) + "개");
+            results.add(parseRank(rank) + result.getCount(rank) + "개");
+        }
+        return results;
+    }
+    private static List<String> parseResult(final LottoGameResult gameResult) {
+        List<String> results = new ArrayList<>();
+
+        for (Rank rank : Rank.reverseValues()) {
+            results.add(parseRank(rank) + gameResult.getRankCount(rank) + "개");
         }
         return results;
     }
 
-    private static String showRank(final Rank rank) {
+    private static String parseRank(final Rank rank) {
         StringBuilder sb = new StringBuilder();
         sb.append(rank.getMatchCount() + "개 일치");
         if (rank == Rank.SECOND) {
